@@ -13,37 +13,32 @@ readme = "README-public.md"
 keywords = [
     "eccenca Corporate Memory", "plugin", "SHACL"
 ]
+homepage = "https://github.com/eccenca/cmem-plugin-pyshacl"
 
 [tool.poetry.dependencies]
-python = "^3.9"
-cmem-plugin-base = ">3.0.0,<4.0.0"
-pyshacl = "^0.22.2"
-validators = "^0.20.0"
-requests-toolbelt = "^0.10.1"
-rdflib = "^6.2.0"
-cmem-cmempy = "^23.1"
-types-setuptools = "^65.5.0.3"
+# if you need to change python version here, change it also in .python-version
+python = "^3.11"
+pyshacl = "^0.26.0"
+validators = "^0.28.1"
+rdflib = "^6.3.2"
 distutils-strtobool = "^0.1.0"
-ruamel-yaml = "^0.17.21"
-PyYAML = "^6.0"
-charset-normalizer = "^3.1.0"
-packaging = "^23.0"
+ruamel-yaml = "^0.18.6"
+
+[tool.poetry.dependencies.cmem-plugin-base]
+version = "^4.5.0"
+allow-prereleases = false
 
 [tool.poetry.group.dev.dependencies]
-bandit = "^1.7.5"
-black = "^23.3.0"
-coverage = "^7.2.3"
-defusedxml = "^0.7.1"
-flake8-formatter-junit-xml = "^0.0.6"
-genbadge = "^1.1.0"
-mypy = "^1.2.0"
-pylint-junit = "^0.3.2"
-pytest = "^7.3.1"
-pytest-cov = "^4.0.0"
-pytest-memray = "^1.4.0"
+genbadge = {extras = ["coverage"], version = "^1.1.1"}
+mypy = "^1.10.0"
+pip = "^24"
+pytest = "^7.4.4"
+pytest-cov = "^4.1.0"
+pytest-dotenv = "^0.5.2"
+pytest-html = "^4.1.1"
+pytest-memray = { version = "^1.6.0",  markers = "platform_system != 'Windows'" }
+ruff = "^0.4.4"
 safety = "^1.10.3"
-typed-ast = "^1.5.4"
-wheel = "^0.38.4"
 
 [build-system]
 requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
@@ -53,21 +48,7 @@ build-backend = "poetry_dynamic_versioning.backend"
 enable = true
 vcs = "git"
 dirty = true
-
-[tool.pylint.MASTER]
-load-plugins="pylint_junit"
-
-[tool.pylint.General]
-ignore = "version.py"
-
-[tool.pylint.'MESSAGES CONTROL']
-extension-pkg-whitelist = "pydantic"
-disable = "fixme"
-max-locals = 17
-
-# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#pylint
-[tool.pylint.messages_control]
-disable = "R0903"
+bump = true
 
 [tool.mypy]
 warn_return_any = true
@@ -75,4 +56,45 @@ ignore_missing_imports = true
 
 [tool.pytest.ini_options]
 addopts = ""
+
+[tool.coverage.report]
+exclude_also = [
+    "def __repr__",
+    "if self.debug:",
+    "if settings.DEBUG",
+    "raise AssertionError",
+    "raise NotImplementedError",
+    "if TYPE_CHECKING:",
+    "class .*\\bProtocol\\):",
+    "@(abc\\.)?abstractmethod",
+    ]
+
+[tool.ruff]
+line-length = 100
+target-version = "py311"
+
+[tool.ruff.format]
+line-ending = "lf" # Use `\n` line endings for all files
+
+[tool.ruff.lint]
+select = ["ALL"]
+ignore = [
+    "ANN101",  # Missing type annotation for self in method
+    "ANN204",  # Missing return type annotation for special method `__init__`
+    "COM812",  # missing-trailing-comma
+    "D107",    # Missing docstring in __init__
+    "D203",    # [*] 1 blank line required before class docstring
+    "D211",    # No blank lines allowed before class docstring
+    "D213",    # Multi-line docstring summary should start at the second line
+    "D400",    # First line should end with a period
+    "D415",    # First line should end with a period, question mark, or exclamation point
+    "EM",      # Exception texts - https://docs.astral.sh/ruff/rules/#flake8-errmsg-em
+    "FBT",     # The Boolean Trap - https://docs.astral.sh/ruff/rules/#flake8-boolean-trap-fbt
+    "FIX002",  # Allow to add TODO notes in the code
+    "G004",    # Logging statement uses f-string
+    "ISC001",  # single-line-implicit-string-concatenation
+    "PD",      # opinionated linting for pandas code
+    "S101",    # use of assert detected
+    "TRY003",  # Avoid specifying long messages outside the exception class
+]
 
