@@ -1,4 +1,1 @@
-"""
-__init.py___
-"""
-__version__ = "4.1.0"
+"""pyshacl - main package"""
