@@ -1,5 +1,7 @@
 """Testing utilities."""
+
 import os
+from typing import ClassVar
 
 import pytest
 
@@ -7,11 +9,11 @@ import pytest
 from cmem.cmempy.api import get_token
 from cmem.cmempy.config import get_oauth_default_credentials
 from cmem_plugin_base.dataintegration.context import (
-    PluginContext,
-    UserContext,
-    TaskContext,
     ExecutionContext,
+    PluginContext,
     ReportContext,
+    TaskContext,
+    UserContext,
 )
 
 needs_cmem = pytest.mark.skipif(
@@ -23,7 +25,7 @@ class TestUserContext(UserContext):
     """dummy user context that can be used in tests"""
 
     __test__ = False
-    default_credential: dict = {}
+    default_credential: ClassVar[dict] = {}
 
     def __init__(self):
         # get access token from default service account
